Fix gcc warning -Wsign-compare

+++ src/mat5.c
@@ -4074,7 +4074,7 @@ GetDataSlab(void *data_in, void *data_out, enum matio_classes class_type,
             err = 1;
         else if ( (size_t)stride[1]*(edge[1]-1)+start[1]+1 > dims[1] )
             err = 1;
-        else if ( ( stride[0] == 1 && edge[0] == dims[0] ) &&
+        else if ( ( stride[0] == 1 && (size_t)edge[0] == dims[0] ) &&
                   ( stride[1] == 1 ) && ( same_type == 1 ) )
             memcpy(data_out, data_in, nbytes);
         else {
